third_party/blink/web_tests/fast/hidpi/image-srcset-src-selection-1x-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS document.getElementById("foo").clientWidth==100 is true
This test passes if the div below is a blue 100px square when the deviceScaleFactor is 1. It simply ensures that the src attribute is taken into account by the selection algorithm when this one is processing the images candidates
